Private messaging system (part 1)
I need a break from coding a messaging system for a while, as I got stuck here and I've leaved a project for few weeks. This update shouldn't be used, as it contains a lot of bugs and not finished features. Also it's not refreshing a content by itself.

WARNING! This update requires to make a database changes manually!

I just don't know what I have changed. Also, I've switched from UniServerZ to WAMPServer, so dbschema.sql is a bit different.