Add an CLI script for recalculating users statistics.
It was required to make this file, in case of a future changes to actions points values or invalid points counting due to a bug. It looks like it's finally counting values correctly.

You can only access this script from a command-line. To do that, in CMD/Terminal go to the "cli/" directory and run: php nameofafile.php.

WARNING! This update requires to make a database changes manually!

ALTER TABLE `posts` ADD `delete_moderator` BOOLEAN NOT NULL DEFAULT FALSE AFTER `status`;