Display a datetime instead of interval on messages.
Now you can see when exactly a message has been sent instead of seeing for example: 14 days ago.  Same string as this could appear a few times in a row. It's displayed in a MySQL datetime format and UTC timezone.

Note, that it's a temporary change and I'll try to make it better soon.