Remember a page that guest was trying to access.
When guest access a page that require a logged user, he/she gets redirected to a login page and previous page link is stored in a session. When user will log in, he/she gets redirected to this page and path is cleared from a session.

This update contains also some bug fixes.