Display a year of starting a website.
You can now add a year to a version.dev.ini file that will be shown in a few places. If you have started in a current year, only this year will be shown, if not, your year and current year will be shown. For example:

Website Title © 2017 - 2018 or Website Title © 2018