Send execution time in JSON if enableDebug is set to true.
enableDebug - can be set to true, as it will be used for displaying some details about server (for now only execution time of scripts).

enableErrors - should be set to false, as it is forcing PHP to return errors. It can leak sensitive informations, I guess. But there's no custom HTTP 500 page yet, so on errors it can display a default browser error for it.